seq v3 (WIP): deleted autoinc from vindexes
