+++ go/vt/vtgate/vindexes/lookup_hash_unique_test.go
@@ -72,13 +72,6 @@ func TestLookupHashUniqueCreate(t *testing.T) {
 	}
 }
 
-func TestLookupHashUniqueGenerate(t *testing.T) {
-	_, ok := lhu.(planbuilder.LookupGenerator)
-	if ok {
-		t.Errorf("lhu.(planbuilder.LookupGenerator): true, want false")
-	}
-}
-
 func TestLookupHashUniqueReverse(t *testing.T) {
 	_, ok := lhu.(planbuilder.Reversible)
 	if ok {
